Move to splinefill.h all the prototypes of non-static splinefill.c functions
