+++ fontforge/Makefile.am
@@ -77,6 +77,7 @@ noinst_HEADERS = \
 	scstyles.h \
 	sfd.h \
 	spiro.h \
+	splinefill.h \
 	fontforgeui.h sftextfieldP.h configure-fontforge.h views.h 	\
 	flaglist.h strlist.h charview_private.h cvundoes.h tables.h
