+++ fontforge/autohint.c
@@ -34,6 +34,7 @@
 #include <stdio.h>
 #include <math.h>
 #include "splinefont.h"
+#include "splinefill.h"
 #include "views.h"
 #include "stemdb.h"
 #include <utype.h>
