+++ fontforge/bitmapchar.c
@@ -32,6 +32,7 @@
 #include "dumpbdf.h"
 #include "fontforgevw.h"
 #include "fvfonts.h"
+#include "splinefill.h"
 #include <string.h>
 #include <ustring.h>
 #include <utype.h>
