+++ fontforge/cvexport.c
@@ -32,6 +32,7 @@
 #include "dumppfa.h"
 #include "fontforgevw.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include <math.h>
 #include <locale.h>
 #include <string.h>
