+++ fontforge/cvundoes.c
@@ -38,6 +38,7 @@
 #include "namelist.h"
 #include "sfd.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include "views.h"
 #include <math.h>
 #include <ustring.h>
