+++ fontforge/dumpbdf.c
@@ -32,6 +32,7 @@
 #include "encoding.h"
 #include "fontforge.h"
 #include "splinefont.h"
+#include "splinefill.h"
 #include <gdraw.h>			/* for the defn of GClut for greymaps */
 #include <string.h>
 #include <ustring.h>
