+++ fontforge/edgelist.h
@@ -78,12 +78,6 @@ typedef struct edgelist {
     DBounds bbox;		/* Not always set. {m,o}{min,max} a provide scaled bbox, this is in glyph units */
 } EdgeList;
 
-extern void FreeEdges(EdgeList *es);
-extern bigreal TOfNextMajor(Edge *e, EdgeList *es, bigreal sought_y );
-extern void FindEdgesSplineSet(SplinePointList *spl, EdgeList *es, int ignore_clip);
-extern Edge *ActiveEdgesInsertNew(EdgeList *es, Edge *active,int i);
-extern Edge *ActiveEdgesRefigure(EdgeList *es, Edge *active,real i);
-extern Edge *ActiveEdgesFindStem(Edge *apt, Edge **prev, real i);
 
 /* Version which is better for everything other than rasterization */
 /*  (I think) */
