+++ fontforge/encoding.c
@@ -35,6 +35,7 @@
 #include "namelist.h"
 #include "psread.h"
 #include "splinefont.h"
+#include "splinefill.h"
 #include <ustring.h>
 #include <utype.h>
 #include <math.h>
