+++ fontforge/fvcomposite.c
@@ -33,6 +33,7 @@
 #include "encoding.h"
 #include "fontforgevw.h"
 #include "fvfonts.h"
+#include "splinefill.h"
 #include "namelist.h"
 #include <chardata.h>
 #include <math.h>
