+++ fontforge/fvfonts.c
@@ -31,6 +31,7 @@
 #include "fontforgevw.h"
 #include "lookups.h"
 #include "namelist.h"
+#include "splinefill.h"
 #include "ustring.h"
 #include "utype.h"
 #include "gfile.h"
