+++ fontforge/macbinary.c
@@ -38,6 +38,7 @@
 #include "mem.h"
 #include "parsepfa.h"
 #include "parsettf.h"
+#include "splinefill.h"
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
