+++ fontforge/palmfonts.c
@@ -32,6 +32,7 @@
 #include "fontforgevw.h"
 #include "macbinary.h"
 #include "mem.h"
+#include "splinefill.h"
 #include <stdio.h>
 #include <math.h>
 #include "splinefont.h"
