+++ fontforge/parsettfbmf.c
@@ -32,6 +32,7 @@
 #include "fontforge.h"
 #include "chardata.h"
 #include "mem.h"
+#include "splinefill.h"
 #include "utype.h"
 #include "ustring.h"
 #include <math.h>
