+++ fontforge/scripting.c
@@ -64,6 +64,7 @@
 #include "search.h"
 #include "sfd.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include <gfile.h>
 #include <utype.h>
 #include <ustring.h>
