+++ fontforge/sfd.c
@@ -41,6 +41,7 @@
 #include "parsettf.h"
 #include "psread.h"
 #include "splinefont.h"
+#include "splinefill.h"
 #include "baseviews.h"
 #include "views.h"
 #include <gdraw.h>
