+++ fontforge/splinefill.c
@@ -24,6 +24,8 @@
  * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+#include "splinefill.h"
+
 #include "fontforge.h"
 #include "fvfonts.h"
 #include "psread.h"
