+++ fontforge/splinefont.c
@@ -42,6 +42,7 @@
 #include "parsepfa.h"
 #include "parsettf.h"
 #include "sfd.h"
+#include "splinefill.h"
 #include <utype.h>
 #include <ustring.h>
 #include <math.h>
