+++ fontforge/splineutil.c
@@ -38,6 +38,7 @@
 #include <math.h>
 #include "psfont.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include "ustring.h"
 #include "utype.h"
 #include "views.h"		/* for FindSel structure */
