+++ fontforge/splineutil2.c
@@ -31,6 +31,7 @@
 #include "namelist.h"
 #include "psread.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include <math.h>
 #include "ustring.h"
 #include "chardata.h"
