+++ fontforgeexe/anchorsaway.c
@@ -26,6 +26,7 @@
  */
 #include "fontforgeui.h"
 #include "fvfonts.h"
+#include "splinefill.h"
 #include <gkeysym.h>
 #include <string.h>
 #include <ustring.h>
