+++ fontforgeexe/charinfo.c
@@ -33,6 +33,7 @@
 #include "fvfonts.h"
 #include "lookups.h"
 #include "namelist.h"
+#include "splinefill.h"
 #include <ustring.h>
 #include <math.h>
 #include <utype.h>
