+++ fontforgeexe/cvimportdlg.c
@@ -30,6 +30,7 @@
 #include "fontforgeui.h"
 #include "fvimportbdf.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include <math.h>
 #include <sys/types.h>
 #include <dirent.h>
