+++ fontforgeexe/cvpalettes.c
@@ -27,6 +27,7 @@
 #include "cvundoes.h"
 #include "fontforgeui.h"
 #include "spiro.h"
+#include "splinefill.h"
 #include "collabclientui.h"
 
 int palettes_docked=1;
