+++ fontforgeexe/savefontdlg.c
@@ -30,6 +30,7 @@
 #include "macbinary.h"
 #include "mm.h"
 #include "namelist.h"
+#include "splinefill.h"
 #include <ustring.h>
 #include <locale.h>
 #include <gfile.h>
