+++ fontforgeexe/showatt.c
@@ -29,6 +29,7 @@
 #include "fvfonts.h"
 #include "glyphcomp.h"
 #include "lookups.h"
+#include "splinefill.h"
 #include <utype.h>
 #include <ustring.h>
 #include <gkeysym.h>
